estates/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.4
django-filter==21.1
django-import-export==2.8.0
django-simple-history==3.1.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
gunicorn==20.0.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
paytmchecksum==1.7.0
Pillow==9.1.1
pycryptodome==3.9.8
pytz==2020.1
PyYAML==6.0
sentry-sdk==1.5.12
sqlparse==0.3.1
tablib==3.2.1
urllib3==1.26.9
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
